requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2
django-cors-headers==4.4.0
django-elasticsearch-dsl==8.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-database-url==2.2.0
elastic-transport==8.15.0
elasticsearch==8.15.1
elasticsearch-dsl==8.15.2
elasticsearch6==6.8.2
kombu==5.4.0
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
